+++ ICD.Connect.Conferencing.Zoom/Components/ZoomRoomComponentFactory.cs
@@ -12,6 +12,7 @@ using ICD.Connect.Conferencing.Zoom.Components.Directory;
 using ICD.Connect.Conferencing.Zoom.Components.Layout;
 using ICD.Connect.Conferencing.Zoom.Components.Presentation;
 using ICD.Connect.Conferencing.Zoom.Components.System;
+using ICD.Connect.Conferencing.Zoom.Components.TraditionalCall;
 
 namespace ICD.Connect.Conferencing.Zoom.Components
 {
@@ -30,7 +31,8 @@ namespace ICD.Connect.Conferencing.Zoom.Components
 				{typeof(BookingsComponent), zoomRoom => new BookingsComponent(zoomRoom)},
 				{typeof(CallComponent), zoomRoom => new CallComponent(zoomRoom)},
 				{typeof(LayoutComponent), zoomRoom => new LayoutComponent(zoomRoom)},
-				{typeof(AudioComponent), zoomRoom => new AudioComponent(zoomRoom)}
+				{typeof(AudioComponent), zoomRoom => new AudioComponent(zoomRoom)},
+				{typeof(TraditionalCallComponent), zoomRoom => new TraditionalCallComponent(zoomRoom)}
 			};
 
 		private readonly ZoomRoom m_ZoomRoom;